requirements.txt
PySimpleGUI
PyPDF2
pdf2image
yagmail
Pillow
